Made a few detection methods functional.
